+++ packages/features/routing-trace/constants.ts
@@ -1,0 +1,9 @@
+export const ROUTING_TRACE_DOMAINS = {
+  SALESFORCE: "salesforce",
+  ROUTING_FORM: "routing_form",
+} as const;
+
+export const ROUTING_TRACE_STEPS = {
+  SALESFORCE_ASSIGNMENT: "salesforce_assignment",
+  ATTRIBUTE_LOGIC_EVALUATED: "attribute-logic-evaluated",
+} as const;
